rename to course block and fix line height for styling
