--- src/views/components/common/CalendarCourseMeeting/CalendarCourseMeeting.module.scss
@@ -1,11 +1,0 @@
-.calendar-course {
-    display: flex;
-    padding: 7px 7px 9px 7px;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 5px;
-    flex: 1 0 0;
-    align-self: stretch;
-    border-radius: 4px;
-    background: #cbd5e1;
-}
